% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animalhealth}
\alias{animalhealth}
\title{Simulated animal health data (MCAR)}
\format{
A 200 x 2 data frame:
\describe{
\item{animal}{The animal of interest}
\item{hear_rate}{The corresponding heart rate of the animal (bpm)}
}
}
\usage{
animalhealth
}
\description{
A toy dataset with heart rate data for various animals.
}
\keyword{datasets}
