% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{drop_na}
\alias{drop_na}
\title{Drops rows containing NAs}
\usage{
drop_na(.data, ...)
}
\arguments{
\item{.data}{a data.frame, data.table, matrix, or list}

\item{...}{Specification of columns to fill, leaving blank will select all
columns. Use bare variable names. Select all variables between x and z with 
x:z, exclude y with -y. For more options, see the \code{dplyr::select} 
documentation. For lists, elements are treated as columns.}

\item{.dots}{Use \code{drop_na_()} to do standard evaluation. See dplyr's 
NSE vignette for more information with \code{vignette("nse", "dplyr")}.}
}
\value{
an object of the same class as \code{.data} with the NA rows dropped
}
\description{
Identical to \code{data[complete.cases(data)]} or 
\code{dplyr::filter(complete.cases(.))}.
}

