% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_cols.R
\name{drop_na_cols}
\alias{drop_na_cols}
\title{Drops columns that contain a percent NAs above a tolerance level}
\usage{
drop_na_cols(.data, tolerance = 0, ...)
}
\arguments{
\item{.data}{a data.frame, matrix, or list}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables. Note: this is the same
behavior as \code{dplyr::select}, in fact the \code{dplyr::select_vars}
function is used here to implement this functionality.}

\item{.dots}{Use \code{drop_na_cols_()} to do standard evaluation.}
}
\value{
an object without columns/elements containing too many NAs
}
\description{
Drops columns that contain a percent NAs above a tolerance level
}
\examples{

drop_na_cols(mtcars)
}

