% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\alias{fill_na_mean}
\alias{fill_na_mean_}
\alias{fill_na_median}
\alias{fill_na_mode}
\title{Fills in (replaces) NAs with non-NA values}
\usage{
fill_na(.data, .fill, ..., .args = NULL, .warning = TRUE,
  .inplace = FALSE)

fill_na_mean(.data, ..., .args = NULL, .warning = TRUE, .inplace = FALSE)

fill_na_mean_(.data, ..., .dots, .args = NULL, .warning = TRUE,
  .inplace = FALSE)

fill_na_median(.data, ..., .args = NULL, .warning = TRUE,
  .inplace = FALSE)

fill_na_mode(.data, ..., .dots, .args = NULL, .warning = TRUE,
  .inplace = FALSE)
}
\arguments{
\item{.data}{a data.frame, data.table, matrix, list, or vector}

\item{.fill}{x}

\item{...}{a selection of columns to fill when \code{.data} is a data.frame, 
data.table, matrix, or list, or arguments to \code{.fill} when \code{.data}
is a vector. 

    If \code{.data} is a data.frame, data.table, matrix, or list, a 
    specification of columns to fill, leaving blank will select all
    columns. Use bare variable names. Select all variables between x and z 
    with x:z, exclude y with -y. For more options, see the 
    \code{dplyr::select} documentation. For lists, elements are treated as 
    columns.
    
    If \code{.data} is a vector, \code{...} are arguments to pass to 
    \code{.fill}. You can still use \code{.args} to supply arguments to 
    \code{.fill}, but you cannot use both \code{...} and \code{.args} at
    the same time.}

\item{.args}{for non-vector inputs, a list of arguments to \code{.fill} if
\code{.fill} is a function (note that the first argument to \code{.fill}
is \code{.data})}

\item{.warning}{\code{TRUE} by default, display a warning when the type of 
\code{.fill} is different from the type of \code{.data}}

\item{.inplace}{\code{FALSE} by default, if \code{.data} is a 
\code{data.table}, fill in the NAs in place (modify \code{.data} by 
reference) if this argument is \code{TRUE}}

\item{.dots}{Use \code{fill_na_()} to do standard evaluation. See dplyr's 
NSE vignette for more information with \code{vignette("nse", "dplyr")}.}
}
\value{
an object of the same class as \code{.data} with the NAs filled in
}
\description{
Fills in (replaces) NAs with non-NA values
}

