% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthcheck}
\alias{healthcheck}
\title{Simulated health check data (MAR)}
\format{
A 200 x 5 data frame:
\describe{
\item{bone_mass}{Bone mass of individual (kg)}
\item{body_fat}{Body fat percentage of individual}
\item{height}{Height of individual (cm)}
\item{age}{Age of individual}
\item{rbc}{Red blood cell count of individual (million/mm^3)}
}
}
\usage{
healthcheck
}
\description{
A toy dataset with typical health check-up metrics for various individuals.
}
\keyword{datasets}
