% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_type}
\alias{identify_type}
\title{Identifies the type of a vector}
\usage{
identify_type(vector)
}
\arguments{
\item{vector}{- a vector to identify the type of}
}
\description{
\code{identify_type} identifies what type a vector is. Generally it agrees
with \code{typeof}, except in the case of factors.
}
\examples{

identify_type(mtcars$mpg)
sapply(iris, identify_type)
}

