% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na}
\alias{na}
\title{Gives summary information on NA values}
\usage{
na(.data, ...)
}
\arguments{
\item{.data}{a data.frame, data.table, matrix, list, or vector}

\item{...}{Specification of columns to fill, leaving blank will select all
columns. Use bare variable names. Select all variables between x and z with 
x:z, exclude y with -y. For more options, see the \code{dplyr::select} 
documentation. For lists, elements are treated as columns.}

\item{.dots}{Use \code{na_()} to do standard evaluation. See dplyr's NSE
vignette for more information with \code{vignette("nse", "dplyr")}.}
}
\value{
a table detailing the NAs in an object
}
\description{
Given an object, prints out a summary table describing the number and 
percent of missing data (NAs), both overall and by column (for data.frames, 
matricies) or element (for lists).
}
\examples{

Examining the NAs in a data.frame
na(mtcars)

The \\code{na} function also works on vectors:
na(LETTERS)

Matricies:
na(as.matrix(mtcars))

and Lists:
na(as.list(mtcars))
}

