% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollutionlevels}
\alias{pollutionlevels}
\title{Simulated pollution level data (MCAR)}
\format{
A (200 x 4) \code{data.frame}:
\describe{
\item{light}{Light pollution of settlement (mag/arcsec^2)}
\item{visual}{Visual pollution of settlement (VPI)}
\item{noise}{Noise pollution of settlement (dB)}
\item{air}{Air pollution of settlement (AQI)}
}
}
\usage{
pollutionlevels
}
\description{
A toy dataset with typical pollution level metrics for various settlements.
}
\keyword{datasets}
