% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testscores}
\alias{testscores}
\title{Simulated test scores data}
\format{
A 200 x 2 data frame:
\describe{
\item{id}{The ID of the student}
\item{score}{The student's score in the test}
}
}
\usage{
testscores
}
\description{
A toy dataset with test scores of various students.
}
\keyword{datasets}
