% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcar.R
\name{mcar}
\alias{mcar}
\title{Missing completely at random (MCAR) test}
\usage{
mcar(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A \code{data.frame}:
\item{statistic}{The d^2 statistic}
\item{degrees_freedom}{Degrees of freedom of chi-squared distribution}
\item{p_val}{P-value of the test}
\item{missing_patterns}{Number of missing patterns}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\code{mcar()} performs Little's MCAR test to test for MCAR.
The null hypothesis is that the data is not MCAR.
}
\details{
This function reproduces the d^2 statistic in equation (5) from [1].
This statistic is used to test for MCAR. Comments reference variables
from the paper (in brackets) to improve readability and traceability.
See \code{vignette("background")} for a more thorough description of the
test and definitions of the variables shown (in brackets).
}
\note{
Code is adapted from \code{mcar_test()} from the naniar package
using base R instead of the tidyverse.
}
\examples{
mcar(airquality)

}
\references{
[1] Little RJA. A Test of Missing Completely at Random for
Multivariate Data with Missing Values. Journal of the
American Statistical Association. 1988;83(404):1198-202.
}
