% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missr-package.R
\docType{package}
\name{missr-package}
\alias{missr}
\alias{missr-package}
\title{missr: Classify Missing Data As MCAR, MAR, or MNAR}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package classifies missing data as MCAR, MAR, or MNAR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NoahHellen/missr}
}

}
\author{
\strong{Maintainer}: Noah William Trelawny Hellen \email{noahhellen@gmail.com} [copyright holder]

}
\keyword{internal}
